feat: use custom webpack config file
In NativeScript 6.4.0, a new feature for custom webpack.config.js file will be introduced. The feature allows you to set custom path to the webpack configuration file that CLI will pass to webpack process.
Use the new feature to set a custom webpack configuration and ensure seamless transition when updating NativeScript (and nativescript-dev-webpack) versions.
This way the template should work out of the box with consecutive releases. Gitignore the `webpack.config.js` file, so it will be created by the postinstall script of the `nativescript-dev-webpack` package.